Update update checking/downloading
- Using new update checking system that is much cleaner than the old
one.
- Now following redirects when downloading updates. This is much more
reliable. It used to fail often because of those damn redirects.
- Updated login message system to be less stupid. It's not just a String
anymore. Instead, it is constructed based on the state of the plugin.
- Dev-builds now override config settings regarding update checking and
downloading. Dev-builds will now ALWAYS check for updates and download
them if found. This should prevent people from lingering on dev-builds.
- Now keeping track of build.number, to be used to differentiate between
dev-builds of the same version.