Further modularize the code
- Removed the bigdoors-api and bigdoors-util modules. They have been
combined into a new bigdoors-core module. The old bigdoors-core module
is now the spigot-core module inside the bigdoors-spigot module.
- The bigdoors-core module no longer relies on anything from bukkit.
- Introduced a few new factories for things such as IPWorld (BigDoors'
version of a world), IPLocation (BigDoor's version of a location) etc.
Eventually, these will be used to replace almost all usages of worlds
and locations etc in the code, to move as many classes (especially
doors and their movers) to bigdoors-core (they are currently in
bigdoors-spigot/spigot-core).