Further improve door creation
- The new door creation system is much closer to functional now. BigDoors can now be constructed, even if the command-side of things is a bit of a mess.
What really needs to be done now is create a system to wait for text input (and re-implement the actual tool and timeouts). The text input system needs to be generic enough to allow multiple implementations. It should be able to register a (specific) CommandWaiter but it should also be possible to write other implementations that request it via a sign or an anvil GUI.
Both methods need to support variable and extensible systems, so you can ask for any type of input with any name/title etc.