Add ErrorProne
- Added ErrorProne to the project. This should make it possible to check for a list of types of bugs. However, it seems ErrorProne is very much prone to errors on its own:
  - ErrorProne does not work without fork=true compilation on Java 16+. While this does not prevent use from using it, it does mean compile times are increased by a factor 3 (getting close to a minute!)
  - ErrorProne throws a bunch of errors because it fails to work together with Lombok. Different combinations of Java/ErrorProne/Lombok will cause ErrorProne to error out for different reasons.
As it stands, I am not willing to sacrifice the convenience offered by Lombok or the massive increase in compile times for ErrorProne. Going forward, it may be nice to set up a system that first Delomboks the code and then uses ErrorProne to check the code. This could perhaps be implemented as a separate build profile to run on commits.