Put all generated properties files in zip
- All generated .properties files as generated by the LocalizationGenerator and used by the Localizer are now stored inside a zip file. The file is named as baseName + ".bundle". The ".bundle" extension is used instead of ".zip" because it's better if users don't mess around with the files inside (any modifications will be reset!) and not making it clear that it's a zip will most likely scare off most users.
This is the first step towards implementing an upgrade path + patching system for the localization system. The idea is that the .bundle files are overwritten with every update and users can supply localization files to override the keys inside the .bundle file. These overrides are kept between plugin updates. This makes it possible to update values for existing keys between versions while also ensuring any modifications made by users are kept.