Separate second block removal pass
Removing blocks happens in two passes: On the first pass, all blocks are removed without applying physics. On the second pass, all blocks along the edges of where the structure used to do a postprocessing step to trigger a physics update which should update any blocks that interacted with the structure before it was removed. This was all handled by the `IAnimatedBlockData#deleteOriginalBlock(applyPhysics)` method. There are two problems with this approach:
  1) It results in the pre/post block removal methods of the IAnimatedBlockHook class being called twice for edge blocks.
  2) It's a leaky abstraction.

To resolve the issues mentioned above, `IAnimatedBlockData#deleteOriginalBlock(applyPhysics)` was changed to `IAnimatedBlockData#deleteOriginalBlock()` that only did what it says: Remove the original block however it wants to. `IAnimatedBlockData#postProcessStructureRemoval()` was introduced as a separate post-processing step that is called after all blocks have been removed. The subclasses of IAnimatedBlockData may handle this however they see fit. This postprocessing method has no callbacks into the IAnimatedBlockHook class. The reasoning for this is that;
  1) I cannot think of any reason why a hook would interact with the physics-updating step of the removal process that cannot be handled in either the regular removal or the spawning steps.
  2) This will make it easier to make changes to the two-pass system in the future.